Modified the Series Episode tray to use the UI layout from Template Builder.

+++ AppCMS/src/main/java/com/viewlift/views/customviews/BaseView.java
@@ -9,6 +9,7 @@ import android.view.Gravity;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.FrameLayout;
+import android.widget.Spinner;
 import android.widget.TextView;
 
 import com.viewlift.models.data.appcms.ui.AppCMSUIKeyType;
@@ -580,18 +581,14 @@ public abstract class BaseView extends FrameLayout {
             if (isLandscape(getContext())) {
                 TabletLandscape tabletLandscape = layout.getTabletLandscape();
                 if (tabletLandscape != null) {
-                    if (getViewWidth(tabletLandscape) != -1) {
-                        if (tabletLandscape.getXAxis() != 0f) {
-                            float scaledX = DEVICE_WIDTH * (tabletLandscape.getXAxis() / STANDARD_TABLET_HEIGHT_PX);
-                            lm = Math.round(scaledX);
-                        }
+                    if (tabletLandscape.getXAxis() != 0f) {
+                        float scaledX = DEVICE_WIDTH * (tabletLandscape.getXAxis() / STANDARD_TABLET_HEIGHT_PX);
+                        lm = Math.round(scaledX);
                     }
 
-                    if (getViewHeight(tabletLandscape) != -1) {
-                        if (tabletLandscape.getYAxis() != 0f) {
-                            float scaledY = DEVICE_HEIGHT * (tabletLandscape.getYAxis() / STANDARD_TABLET_WIDTH_PX);
-                            tm = Math.round(scaledY);
-                        }
+                    if (tabletLandscape.getYAxis() != 0f) {
+                        float scaledY = DEVICE_HEIGHT * (tabletLandscape.getYAxis() / STANDARD_TABLET_WIDTH_PX);
+                        tm = Math.round(scaledY);
                     }
 
                     if (getViewWidth(tabletLandscape) == -1f && tabletLandscape.getXAxis() != 0f) {
@@ -647,18 +644,14 @@ public abstract class BaseView extends FrameLayout {
             } else {
                 TabletPortrait tabletPortrait = layout.getTabletPortrait();
                 if (tabletPortrait != null) {
-                    if (getViewWidth(tabletPortrait) != -1) {
-                        if (tabletPortrait.getXAxis() != 0f) {
-                            float scaledX = DEVICE_WIDTH * (tabletPortrait.getXAxis() / STANDARD_TABLET_WIDTH_PX);
-                            lm = Math.round(scaledX);
-                        }
+                    if (tabletPortrait.getXAxis() != 0f) {
+                        float scaledX = DEVICE_WIDTH * (tabletPortrait.getXAxis() / STANDARD_TABLET_WIDTH_PX);
+                        lm = Math.round(scaledX);
                     }
 
-                    if (getViewHeight(tabletPortrait) != -1) {
-                        if (tabletPortrait.getYAxis() != 0f) {
-                            float scaledY = DEVICE_HEIGHT * (tabletPortrait.getYAxis() / STANDARD_TABLET_HEIGHT_PX);
-                            tm = Math.round(scaledY);
-                        }
+                    if (tabletPortrait.getYAxis() != 0f) {
+                        float scaledY = DEVICE_HEIGHT * (tabletPortrait.getYAxis() / STANDARD_TABLET_HEIGHT_PX);
+                        tm = Math.round(scaledY);
                     }
 
                     if (getViewWidth(tabletPortrait) == -1 && tabletPortrait.getXAxis() != 0f) {
@@ -715,18 +708,14 @@ public abstract class BaseView extends FrameLayout {
         } else {
             Mobile mobile = layout.getMobile();
             if (mobile != null) {
-                if (getViewWidth(mobile) != -1) {
-                    if (mobile.getXAxis() != 0f) {
-                        float scaledX = DEVICE_WIDTH * (mobile.getXAxis() / STANDARD_MOBILE_WIDTH_PX);
-                        lm = Math.round(scaledX);
-                    }
+                if (mobile.getXAxis() != 0f) {
+                    float scaledX = DEVICE_WIDTH * (mobile.getXAxis() / STANDARD_MOBILE_WIDTH_PX);
+                    lm = Math.round(scaledX);
                 }
 
-                if (getViewHeight(mobile) != -1) {
-                    if (mobile.getYAxis() != 0f) {
-                        float scaledY = DEVICE_HEIGHT * (mobile.getYAxis() / STANDARD_MOBILE_HEIGHT_PX);
-                        tm = Math.round(scaledY);
-                    }
+                if (mobile.getYAxis() != 0f) {
+                    float scaledY = DEVICE_HEIGHT * (mobile.getYAxis() / STANDARD_MOBILE_HEIGHT_PX);
+                    tm = Math.round(scaledY);
                 }
 
                 if (getViewWidth(mobile) == -1 && mobile.getXAxis() != 0f) {
@@ -797,9 +786,15 @@ public abstract class BaseView extends FrameLayout {
                 componentKey = AppCMSUIKeyType.PAGE_EMPTY_KEY;
             }
 
-            if (jsonValueKeyMap.get(viewType) == AppCMSUIKeyType.PAGE_SEASON_TRAY_MODULE_KEY) {
+            if (jsonValueKeyMap.get(viewType) == AppCMSUIKeyType.PAGE_SEASON_TRAY_MODULE_KEY &&
+                    view instanceof Spinner) {
                 viewHeight = LayoutParams.WRAP_CONTENT;
                 viewWidth = LayoutParams.WRAP_CONTENT;
+            }
+
+            AppCMSUIKeyType componentViewType = jsonValueKeyMap.get(viewType);
+            if (componentViewType == null) {
+                componentViewType = AppCMSUIKeyType.PAGE_EMPTY_KEY;
             }
 
             switch (componentKey) {
@@ -813,16 +808,19 @@ public abstract class BaseView extends FrameLayout {
                     break;
 
                 case PAGE_VIDEO_PLAY_BUTTON_KEY:
-                    lm -= 8;
-                    rm -= 8;
-                    bm -= 8;
-                    tm -= 8;
+                    if (jsonValueKeyMap.get(viewType) != AppCMSUIKeyType.PAGE_SEASON_TRAY_MODULE_KEY) {
+                        lm -= 8;
+                        rm -= 8;
+                        bm -= 8;
+                        tm -= 8;
+                    }
                     break;
 
                 case PAGE_PLAY_IMAGE_KEY:
                     if (AppCMSUIKeyType.PAGE_HISTORY_MODULE_KEY != jsonValueKeyMap.get(viewType)
                             && AppCMSUIKeyType.PAGE_DOWNLOAD_MODULE_KEY != jsonValueKeyMap.get(viewType)
-                            && AppCMSUIKeyType.PAGE_WATCHLIST_MODULE_KEY != jsonValueKeyMap.get(viewType)) {
+                            && AppCMSUIKeyType.PAGE_WATCHLIST_MODULE_KEY != jsonValueKeyMap.get(viewType)
+                            && componentViewType != AppCMSUIKeyType.PAGE_SEASON_TRAY_MODULE_KEY) {
                         gravity = Gravity.CENTER;
                         lm -= 40;
                         rm -= 40;
@@ -945,18 +943,6 @@ public abstract class BaseView extends FrameLayout {
                 componentType == AppCMSUIKeyType.PAGE_BUTTON_KEY ||
                 componentType == AppCMSUIKeyType.PAGE_IMAGE_KEY) {
             layoutParams.gravity = gravity;
-        }
-
-        if (jsonValueKeyMap.get(viewType) == AppCMSUIKeyType.PAGE_SEASON_TRAY_MODULE_KEY) {
-            float scaledY = convertVerticalValue(getContext(),
-                    getYAxis(getContext(), parentLayout, 0.0f));
-            if (componentType == AppCMSUIKeyType.PAGE_COLLECTIONGRID_KEY) {
-                tm = (int) (convertVerticalValue(getContext(),
-                        getYAxis(getContext(), layout, 0.0f)));
-            }
-            tm += Math.round(scaledY);
-            layoutParams.setMargins(lm, tm, rm, bm);
-            layoutParams.height = LayoutParams.WRAP_CONTENT;
         }
 
         if (componentType == AppCMSUIKeyType.PAGE_COLLECTIONGRID_KEY) {
